refactor: map struct mapper config 수정
