--- src/main/java/net/devgrr/interp/ia/api/config/mapStruct/IssueMapperConfig.java
@@ -1,13 +1,0 @@
-package net.devgrr.interp.ia.api.config.mapStruct;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class IssueMapperConfig {
-
-  @Bean
-  public IssueMapper issueMapper() {
-    return new IssueMapperImpl();
-  }
-}
