--- src/main/java/net/devgrr/interp/ia/api/config/mapStruct/MemberMapperConfig.java
@@ -1,12 +1,0 @@
-package net.devgrr.interp.ia.api.config.mapStruct;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class MemberMapperConfig {
-  @Bean
-  public MemberMapper memberMapper() {
-    return new MemberMapperImpl();
-  }
-}
